fixed #563

+++ cocos2dx/support/ccUtils.cpp
@@ -25,7 +25,7 @@ THE SOFTWARE.
 
 namespace cocos2d
 {
-	unsigned int ccNextPOT(unsigned int x)
+	unsigned long ccNextPOT(unsigned long x)
 	{
 		x = x - 1;
 		x = x | (x >> 1);
